Saving loggedIn state, and some changes
Created system to check if user is logged in or not.
Added conditional to check for this state when index.html is accessed, redirecting to login.html in case there's no loggedIn state on localStorage.
Added extra link on index.html's navbar, showing an option to log out, stylizing said element on styles.css.
Added id and autocomplete attributes to input boxes, for better compatibility.